--- core/src/main/scala/tech/beshu/ror/proxy/RorProxy.scala
@@ -1,39 +1,0 @@
-/*
- *    This file is part of ReadonlyREST.
- *
- *    ReadonlyREST is free software: you can redistribute it and/or modify
- *    it under the terms of the GNU General Public License as published by
- *    the Free Software Foundation, either version 3 of the License, or
- *    (at your option) any later version.
- *
- *    ReadonlyREST is distributed in the hope that it will be useful,
- *    but WITHOUT ANY WARRANTY; without even the implied warranty of
- *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- *    GNU General Public License for more details.
- *
- *    You should have received a copy of the GNU General Public License
- *    along with ReadonlyREST.  If not, see http://www.gnu.org/licenses/
- */
-package tech.beshu.ror.proxy
-
-import better.files.File
-import cats.effect.IO
-import tech.beshu.ror.boot.StartingFailure
-import tech.beshu.ror.proxy.RorProxy.CloseHandler
-
-trait RorProxy {
-
-  def config: RorProxy.Config
-
-  def start: IO[Either[StartingFailure, CloseHandler]]
-
-}
-
-object RorProxy {
-  type CloseHandler = () => IO[Unit]
-
-  final case class Config(targetEsNode: String,
-                          proxyPort: String,
-                          esConfigFile: File)
-
-}
